Remove any client reference
Will enable us to focus on getting the server stable without having to worry about client.
Will come back to work on client later, probably in another branch, as to be able to experiment with it.

This change is also nice in a refactoring point of view, since it will enable us to refactor a lot of the current code