Require handlers to return *Response directly
This commit changes the generated handler methods inside ServerInterface
to directly return a *Response value instead of requiring it to call
`render.Render` manually. This results in much cleaner handler code.

While this change breaks all existing code, it is not a significant
change by any means, so most changes can be regexed away to work.