Implement x-go-optional-value, x-go-string (#82)
* Implement x-go-optional-value, x-go-string

This commit implements x-go-optional-value, which forces the code
generator to generate value types (not pointer types) for optional
parameters. This closes issue #76.

This commit also implements x-go-string, which adds the ",string" suffix
into the JSON tag of a field, allowing that field to be parsed from a
string to an integer of any type. This closes issue #77.

* Remove extParse{OptionalValue,OmitEmpty}

The removed functions are redundant when extParseBool now exists.

* Document x-go-{optional-value,string,omitempty}